# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: LyoDex API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /requests:
    get:
      operationId: listRequests
      tags: [requests]
      summary: List all freeze-dry requests
      responses:
        "200":
          description: List of requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FreezeDryRequest"
    post:
      operationId: createRequest
      tags: [requests]
      summary: Submit a new freeze-dry request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FreezeDryRequestInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FreezeDryRequest"
  /requests/{id}:
    get:
      operationId: getRequest
      tags: [requests]
      summary: Get a request by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Request found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FreezeDryRequest"
        "404":
          description: Not found
  /requests/{id}/messages:
    get:
      operationId: listMessagesForRequest
      tags: [requests]
      summary: List messages for a request
      description: Returns the message thread for a request. Only the buyer who owns the request, operators who have bid on it, and admins can access this endpoint.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RequestMessage"
        "401":
          description: Unauthenticated
        "403":
          description: Forbidden
        "404":
          description: Not found
    post:
      operationId: postMessageForRequest
      tags: [requests]
      summary: Post a message on a request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestMessageInput"
      responses:
        "201":
          description: Message posted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestMessage"
        "401":
          description: Unauthenticated
        "403":
          description: Forbidden
        "404":
          description: Not found
...
components:
  schemas:
...
    FreezeDryRequest:
      type: object
      required: [id, material_type, quantity_kg, target_moisture, deadline, budget_per_kg, status, created_at]
      properties:
        id:
          type: integer
        material_type:
          type: string
        quantity_kg:
          type: integer
        target_moisture:
          type: ["number", "null"]
        deadline:
          type: string
        budget_per_kg:
          type: ["number", "null"]
        special_requirements:
          type: ["string", "null"]
        status:
          type: string
        buyer_email:
          type: ["string", "null"]
        bid_count:
          type: integer
        created_at:
          type: string
        moderation_note:
          type: ["string", "null"]
        moderated_by:
          type: ["string", "null"]
    FreezeDryRequestInput:
      type: object
      required: [material_type, quantity_kg, deadline]
      properties:
        material_type:
          type: string
        quantity_kg:
          type: integer
        target_moisture:
          type: number
        deadline:
          type: string
        budget_per_kg:
          type: number
        special_requirements:
          type: string
        buyer_email:
          type: string
...
    RequestMessage:
      type: object
      required: [id, request_id, user_id, sender_name, sender_role, body, created_at]
      properties:
        id:
          type: integer
        request_id:
          type: integer
        user_id:
          type: integer
        sender_name:
          type: string
        sender_role:
          type: string
        body:
          type: string
        created_at:
          type: string
    RequestMessageInput:
      type: object
      required: [body]
      properties:
        body:
          type: string
...